Improved background, adds mouse follower
