--- app/components/settings/language.tsx
@@ -1,17 +1,0 @@
-import { useContext } from "react";
-import { SettingsContext } from "./settings";
-import en from "../../resources/languages/en.json";
-import es from "../../resources/languages/es.json";
-
-export function GetLanguage() {
-    const { selectedOption } = useContext(SettingsContext);
-  
-    switch(selectedOption) {
-      case 'Spanish' || 'Español':
-        return es;
-  
-      default:
-      case 'English' || 'Inglés':
-        return en;
-    }
-  }
