+++ app/components/settings/settings.tsx
@@ -2,8 +2,9 @@
 
 import { ReactElement, ReactNode, createContext, useContext, useState } from "react";
 import { GetDropDownMenu } from "../dropdown/dropdown";
-import { GetLanguage } from "./language";
 import styles from "./settings.module.css";
+import en from "../../resources/languages/en.json";
+import es from "../../resources/languages/es.json";
 
 export const SettingsContext = createContext({ isEnabled: false, toggleDisplay: () => {}, selectedOption: '', setSelectedOption: (value: string) => {} });
 
@@ -40,7 +41,7 @@ export function GetSettings(): ReactElement {
     background = json.dialogs.settings.background;
   });
 
-  const { isEnabled, toggleDisplay, selectedOption, setSelectedOption  } = useContext(SettingsContext);
+  const { isEnabled, toggleDisplay, selectedOption, setSelectedOption } = useContext(SettingsContext);
 
   const handleDropdownChange = (selectedOption:string) => {
       setSelectedOption(selectedOption);
@@ -50,6 +51,25 @@ export function GetSettings(): ReactElement {
   <div className={styles.settings}>
     <button className={styles.toggleButton} onClick={toggleDisplay}>{ isEnabled ? disable : enable } {background}</button>
     <GetDropDownMenu onChange={handleDropdownChange} title={title} items={items}/>
+    {IsBackgroundEnabled() && <button className={styles.toggleButton}>Play Game</button>}
   </div>
   );
+}
+
+export function IsBackgroundEnabled(): boolean {
+  const { isEnabled } = useContext(SettingsContext);
+  return isEnabled;
+}
+
+export function GetLanguage() {
+  const { selectedOption } = useContext(SettingsContext);
+
+  switch(selectedOption) {
+    case 'Spanish' || 'Español':
+      return es;
+
+    default:
+    case 'English' || 'Inglés':
+      return en;
+  }
 }
