--- app/components/star-background/star-background.module.css
@@ -1,43 +1,0 @@
-.starBackground {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    pointer-events: none; /* Allows mouse events to pass through the stars */
-}
-
-.star {
-    position: absolute;
-    width: 2px;
-    height: 2px;
-    background-color: white;
-    border-radius: 50%;
-    /* Position the stars randomly */
-    top: calc(100% * var(--randomY));
-    left: calc(100% * var(--randomX));
-}
-
-/* Adjust the size and color of stars here */
-.star:nth-child(1) {
-    width: 4px;
-    height: 4px;
-    background-color: #FFF;
-}
-.star:nth-child(2) {
-    width: 3px;
-    height: 3px;
-    background-color: #DDD;
-}
-/* Add more star styles as needed */
-
-/* Parallax effect */
-.starBackground::before {
-    content: '';
-    position: absolute;
-    top: var(--mouseY);
-    left: var(--mouseX);
-    width: 1px;
-    height: 1px;
-    background-color: transparent;
-}
