+++ app/page.module.css
@@ -14,6 +14,10 @@
   min-height: -webkit-fill-available;
 }
 
+.headerContainer {
+  z-index: 3;
+}
+
 .header {
   display: flex;
   flex-direction: column;
@@ -30,6 +34,7 @@
 .line {
   border-bottom: 2px solid var(--kobe);
   width: 90vw;
+  z-index: 3;
 }
 
 .description {
